+++ webservice/config/config.json
@@ -1,6 +1,6 @@
 {
   "database": {
-    "hostname": "localhost",
+    "hostname": "0.0.0.0",
     "port": 3306,
     "username": "root",
     "password": "",
@@ -9,7 +9,7 @@
   },
 
   "server": {
-    "hostname": "localhost",
+    "hostname": "0.0.0.0",
     "useHTTP": true,
     "useHTTPS": false,
     "HTTPPort": 8088,
