+++ webservice/go.mod
@@ -1,24 +1,12 @@
-module github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments
+module github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/webservice
 
 require (
+	github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments v0.0.0-20190225143407-d4a2c5d1bce5
 	github.com/go-sql-driver/mysql v1.4.1
 	github.com/gorilla/handlers v1.4.0
 	github.com/gorilla/mux v1.7.0
 	github.com/gorilla/sessions v1.1.3
-	github.com/kr/pretty v0.1.0 // indirect
-	github.com/microcosm-cc/bluemonday v1.0.2 // indirect
 	github.com/rs/xid v1.2.1
-	github.com/sergi/go-diff v1.0.0 // indirect
 	github.com/shurcooL/github_flavored_markdown v0.0.0-20181002035957-2122de532470
-	github.com/shurcooL/go v0.0.0-20190121191506-3fef8c783dec // indirect
-	github.com/shurcooL/go-goon v0.0.0-20170922171312-37c2f522c041 // indirect
-	github.com/shurcooL/highlight_diff v0.0.0-20181222201841-111da2e7d480 // indirect
-	github.com/shurcooL/highlight_go v0.0.0-20181215221002-9d8641ddf2e1 // indirect
-	github.com/shurcooL/octicon v0.0.0-20181222203144-9ff1a4cf27f4 // indirect
-	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
-	github.com/sourcegraph/annotate v0.0.0-20160123013949-f4cad6c6324d // indirect
-	github.com/sourcegraph/syntaxhighlight v0.0.0-20170531221838-bd320f5d308e // indirect
-	github.com/stretchr/testify v1.3.0 // indirect
 	golang.org/x/crypto v0.0.0-20190211182817-74369b46fc67
-	google.golang.org/appengine v1.4.0 // indirect
 )
