+++ webservice/model/assignment.go
@@ -1,7 +1,7 @@
 package model
 
 import (
-	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/shared/db"
+	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/webservice/shared/db"
 	_ "github.com/go-sql-driver/mysql" //database driver
 	"time"
 )
