+++ webservice/shared/db/db.go
@@ -4,6 +4,7 @@ import (
 	"database/sql"
 	"fmt"
 	_ "github.com/go-sql-driver/mysql" //database driver
+	"os"
 )
 
 const driverName = "mysql"
@@ -26,7 +27,16 @@ type MySQLInfo struct {
 //ConfigureDB loads database connection string
 func ConfigureDB(info *MySQLInfo) { // TODO (Svein): Is DB needed at the end? We are in db-package
 	// root:@tcp(127.0.0.1:3306)/cs53
-	dataSourceName = fmt.Sprintf("%s:%s@tcp(%s:%d)/%s?parseTime=%t", info.Username, info.Password, info.Hostname, info.Port, info.Database, info.ParseTime)
+	if os.Getenv("DATABASE_URL") == "" {
+		dataSourceName = fmt.Sprintf("%s:%s@tcp(%s:%d)/%s?parseTime=%t", info.Username, info.Password, info.Hostname, info.Port, info.Database, info.ParseTime)
+	} else {
+		dataSourceName = fmt.Sprintf("%s:%s@tcp(%s)/%s?parseTime=%t",
+			os.Getenv("DATABASE_USER"),
+			os.Getenv("DATABASE_PASSWORD"),
+			os.Getenv("DATABASE_URL"),
+			os.Getenv("DATABASE_DATABASE"),
+			info.ParseTime)
+	}
 }
 
 //OpenDB creates connection with database
